Better thread utilization
The current behavior equally distributes work among all worker threads. If one thread is finished faster than others, it would idle while the work queues of other threads might still be full. This PR changes it so that the per-thread queues have at most one element and most jobs are kept in `myPendingJobQueue`, where they are automatically given to the first thread that is idle.